+++ display.h
@@ -9,7 +9,7 @@ extern uint8_t max7219_buffer[MAX7219_BUFFER_SIZE];
 
 void max7219b_clrAll(void);
 void display_init(void);
-void display_set(int x, int y);
+void display_set(uint8_t x, uint8_t y);
 void display_output(void);
 
 #endif
